# [Service#power_state] Avoid update if values match